Try to fix float32 vs float64

+++ sequence/Readers/WeightObjects.py
@@ -29,7 +29,7 @@ def evaluate_object_weights(df, bins_vars, add_syst, name, nuisances):
         mean = wksum / wsum
         unc_up = np.sqrt((wdkupsum / wsum**2) + addsyst**2)
         unc_down = -1.*np.sqrt((wdkdownsum / wsum**2) + addsyst**2)
-        return mean, unc_up, unc_down
+        return mean.astype(np.float32), unc_up.astype(np.float32), unc_down.astype(np.float32)
 
     @cachedmethod(operator.attrgetter('cache'), key=partial(hashkey, 'fevaluate_object_weights'))
     def fevaluate_object_weights(ev, evidx, nsig, source, name_):
